Fix en passant bug
Removed `7 - rank` in en passant field indexing. The number in the "en passant" field in a FEN already refers to chess rank. The `7 - rank` was needed in the main part of this function because `rank` there refers to array index in the order they appear in the FEN, where the first item of the array actually refers to rank 7 (last rank), and so on.